update CSS for index.html and link to recipes
I updated CSS for index.html to style the page the way I wanted it to
look. Had to add a few div elements in the HTML too, along with class
names. I also linked the style sheet to the recipe pages and began to
add CSS styles to them.