rename style.css -> styles.css and update styles
I renamed the style sheet file from style.css to styles.css. Also,
implemented a flex-box layout to the index.html webpage. This also
included some changes made to the actual html on index.html. The recipe
page for "Braised Short Ribs" was also styled and added a link that
will take you back to the homepage at the bottom.